Move docs and typing from stubs to code
