+++ crudclient/testing/response_builder/api_patterns.py
@@ -1,3 +1,10 @@
+"""
+API pattern builder utilities for mock client.
+
+This module provides utilities for creating mock responses that follow common API patterns,
+such as REST resources, nested resources, batch operations, GraphQL endpoints, and OAuth flows.
+"""
+
 from typing import Any, Callable, Dict, List, Optional, TypedDict, Union
 
 from .response import MockResponse
@@ -14,6 +21,13 @@ class PatternDict(TypedDict, total=False):
 
 
 class APIPatternBuilder:
+    """
+    Builder for creating mock responses that follow common API patterns.
+
+    This class provides static methods for generating response patterns that match
+    common API design patterns, making it easier to create realistic mock APIs
+    for testing client code.
+    """
 
     @staticmethod
     def rest_resource(
@@ -28,6 +42,28 @@ class APIPatternBuilder:
         filter_response: Optional[Union[List[Dict[str, Any]], Callable[..., MockResponse]]] = None,
         patch_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
     ) -> List[PatternDict]:
+        """
+        Create response patterns for a standard REST resource.
+
+        This method generates response patterns for standard REST operations
+        (list, get, create, update, delete) on a resource, as well as search
+        and filter operations.
+
+        Args:
+            base_path: The base URL path for the resource (e.g., "users")
+            resource_id_pattern: Regex pattern to match resource IDs in URLs
+            list_response: Response for GET requests to the collection (list operation)
+            get_response: Response for GET requests to a specific resource
+            create_response: Response for POST requests to the collection (create operation)
+            update_response: Response for PUT requests to a specific resource
+            delete_response: Response for DELETE requests to a specific resource
+            search_response: Response for GET requests with a search parameter
+            filter_response: Response for GET requests with filter parameters
+            patch_response: Response for PATCH requests to a specific resource
+
+        Returns:
+            A list of response pattern dictionaries that can be used with a mock client
+        """
         patterns = []
 
         # Ensure base_path starts with / and doesn't end with /
@@ -101,6 +137,26 @@ class APIPatternBuilder:
         update_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
         delete_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
     ) -> List[PatternDict]:
+        """
+        Create response patterns for a nested resource.
+
+        This method generates response patterns for operations on a resource that is
+        nested under a parent resource (e.g., /users/{user_id}/posts/{post_id}).
+
+        Args:
+            parent_path: The path of the parent resource (e.g., "users")
+            child_path: The path of the child resource (e.g., "posts")
+            parent_id_pattern: Regex pattern to match parent resource IDs
+            child_id_pattern: Regex pattern to match child resource IDs
+            list_response: Response for GET requests to the child collection
+            get_response: Response for GET requests to a specific child resource
+            create_response: Response for POST requests to the child collection
+            update_response: Response for PUT requests to a specific child resource
+            delete_response: Response for DELETE requests to a specific child resource
+
+        Returns:
+            A list of response pattern dictionaries that can be used with a mock client
+        """
         patterns = []
 
         # Ensure paths are properly formatted
@@ -137,6 +193,21 @@ class APIPatternBuilder:
         batch_update_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
         batch_delete_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
     ) -> List[PatternDict]:
+        """
+        Create response patterns for batch operations.
+
+        This method generates response patterns for batch operations (create, update, delete)
+        on a resource, which are commonly used for bulk operations.
+
+        Args:
+            base_path: The base URL path for the resource (e.g., "users")
+            batch_create_response: Response for batch create operations
+            batch_update_response: Response for batch update operations
+            batch_delete_response: Response for batch delete operations
+
+        Returns:
+            A list of response pattern dictionaries that can be used with a mock client
+        """
         patterns = []
 
         # Ensure base_path starts with / and doesn't end with /
@@ -170,6 +241,20 @@ class APIPatternBuilder:
         query_matchers: Optional[Dict[str, Union[Dict[str, Any], Callable[..., MockResponse]]]] = None,
         default_response: Optional[Union[Dict[str, Any], Callable[..., MockResponse]]] = None,
     ) -> List[PatternDict]:
+        """
+        Create response patterns for a GraphQL endpoint.
+
+        This method generates response patterns for a GraphQL endpoint, with support
+        for matching specific queries and providing appropriate responses.
+
+        Args:
+            url_pattern: Regex pattern to match the GraphQL endpoint URL
+            query_matchers: Dictionary mapping query patterns to responses
+            default_response: Default response for queries that don't match any pattern
+
+        Returns:
+            A list of response pattern dictionaries that can be used with a mock client
+        """
         patterns = []
 
         # Add patterns for specific queries
@@ -204,6 +289,22 @@ class APIPatternBuilder:
         error_response: Optional[Dict[str, Any]] = None,
         valid_credentials: Optional[Dict[str, str]] = None,
     ) -> List[PatternDict]:
+        """
+        Create response patterns for an OAuth token endpoint.
+
+        This method generates response patterns for an OAuth token endpoint,
+        with support for validating credentials and returning appropriate
+        success or error responses.
+
+        Args:
+            token_url_pattern: Regex pattern to match the token endpoint URL
+            success_response: Response to return for valid credentials
+            error_response: Response to return for invalid credentials
+            valid_credentials: Dictionary of valid credentials to check against
+
+        Returns:
+            A list of response pattern dictionaries that can be used with a mock client
+        """
         if success_response is None:
             success_response = {
                 "access_token": "mock-access-token",
