+++ crudclient/testing/response_builder/validation.py
@@ -4,6 +4,10 @@ from .response import MockResponse
 
 
 class ValidationErrorBuilder:
+    """
+    Provides static methods to create MockResponse objects representing
+    various schema validation errors.
+    """
 
     @staticmethod
     def create_schema_validation_error(
@@ -13,6 +17,19 @@ class ValidationErrorBuilder:
         message: str = "Validation failed",
         error_format: str = "standard",
     ) -> MockResponse:
+        """
+        Creates a MockResponse for a schema validation error.
+
+        Args:
+            invalid_fields: A dictionary mapping field names to error messages.
+            status_code: The HTTP status code for the response.
+            error_code: A custom error code string.
+            message: The main error message.
+            error_format: The format of the error response ('standard', 'json_api', 'detailed', 'simple').
+
+        Returns:
+            A MockResponse object representing the validation error.
+        """
         if error_format == "standard":
             response_data = {
                 "error": error_code,
@@ -56,6 +73,20 @@ class ValidationErrorBuilder:
         error_code: str = "VALIDATION_ERROR",
         error_format: str = "standard",
     ) -> Callable[[Dict[str, Any]], Optional[MockResponse]]:
+        """
+        Creates a validator function for a single field.
+
+        Args:
+            field_name: The name of the field to validate.
+            validators: A list of validator functions for the field.
+            status_code: The status code for validation errors.
+            error_code: The error code for validation errors.
+            error_format: The error format for validation errors.
+
+        Returns:
+            A function that takes request data and returns a MockResponse if validation fails, else None.
+        """
+
         def validator_function(data: Dict[str, Any]) -> Optional[MockResponse]:
             if not isinstance(data, dict):
                 return ValidationErrorBuilder.create_schema_validation_error(
@@ -89,6 +120,21 @@ class ValidationErrorBuilder:
         error_format: str = "standard",
         require_all_fields: bool = False,
     ) -> Callable[[Dict[str, Any]], Optional[MockResponse]]:
+        """
+        Creates a validator function for multiple fields in a data dictionary.
+
+        Args:
+            validators: A dictionary mapping field names to lists of validator functions.
+            status_code: The status code for validation errors.
+            error_code: The error code for validation errors.
+            message: The main error message for validation failures.
+            error_format: The error format for validation errors.
+            require_all_fields: If True, missing fields defined in validators will cause an error.
+
+        Returns:
+            A function that takes request data and returns a MockResponse if validation fails, else None.
+        """
+
         def validator_function(data: Dict[str, Any]) -> Optional[MockResponse]:
             if not isinstance(data, dict):
                 return ValidationErrorBuilder.create_schema_validation_error(
@@ -122,6 +168,10 @@ class ValidationErrorBuilder:
 
 
 class BusinessLogicConstraintBuilder:
+    """
+    Provides static methods to create MockResponse objects representing
+    various business logic constraint violations.
+    """
 
     @staticmethod
     def create_business_rule_error(
@@ -131,6 +181,19 @@ class BusinessLogicConstraintBuilder:
         error_code: str = "BUSINESS_RULE_VIOLATION",
         details: Optional[Dict[str, Any]] = None,
     ) -> MockResponse:
+        """
+        Creates a MockResponse for a generic business rule violation.
+
+        Args:
+            rule_name: The name of the violated rule.
+            message: The error message describing the violation.
+            status_code: The HTTP status code.
+            error_code: The custom error code.
+            details: Optional additional details about the error.
+
+        Returns:
+            A MockResponse object representing the business rule error.
+        """
         response_data: Dict[str, Any] = {"error": error_code, "message": message, "rule": rule_name}
 
         if details:
@@ -146,6 +209,19 @@ class BusinessLogicConstraintBuilder:
         status_code: int = 409,
         error_code: str = "UNIQUE_CONSTRAINT_VIOLATION",
     ) -> MockResponse:
+        """
+        Creates a MockResponse for a unique constraint violation.
+
+        Args:
+            field_name: The name of the field that must be unique.
+            value: The value that caused the violation.
+            entity_type: The type of entity being created/updated.
+            status_code: The HTTP status code (typically 409 Conflict).
+            error_code: The custom error code.
+
+        Returns:
+            A MockResponse object representing the unique constraint error.
+        """
         message = f"A {entity_type} with {field_name} '{value}' already exists"
 
         return MockResponse(status_code=status_code, json_data={"error": error_code, "message": message, "field": field_name, "value": value})
@@ -158,6 +234,19 @@ class BusinessLogicConstraintBuilder:
         status_code: int = 422,
         error_code: str = "FOREIGN_KEY_CONSTRAINT_VIOLATION",
     ) -> MockResponse:
+        """
+        Creates a MockResponse for a foreign key constraint violation.
+
+        Args:
+            field_name: The name of the foreign key field.
+            value: The value provided for the foreign key.
+            referenced_entity: The type of entity being referenced.
+            status_code: The HTTP status code.
+            error_code: The custom error code.
+
+        Returns:
+            A MockResponse object representing the foreign key constraint error.
+        """
         message = f"No {referenced_entity} found with {field_name} '{value}'"
 
         return MockResponse(
@@ -174,6 +263,20 @@ class BusinessLogicConstraintBuilder:
         status_code: int = 422,
         error_code: str = "INVALID_STATE_TRANSITION",
     ) -> MockResponse:
+        """
+        Creates a MockResponse for an invalid state transition error.
+
+        Args:
+            entity_type: The type of entity whose state transition failed.
+            current_state: The current state of the entity.
+            target_state: The attempted target state.
+            allowed_transitions: A list of valid target states from the current state.
+            status_code: The HTTP status code.
+            error_code: The custom error code.
+
+        Returns:
+            A MockResponse object representing the state transition error.
+        """
         message = f"Cannot transition {entity_type} from '{current_state}' to '{target_state}'"
 
         return MockResponse(
@@ -195,6 +298,19 @@ class BusinessLogicConstraintBuilder:
         status_code: int = 422,
         error_code: str = "DEPENDENCY_CONSTRAINT_VIOLATION",
     ) -> MockResponse:
+        """
+        Creates a MockResponse for a dependency constraint violation.
+
+        Args:
+            entity_type: The type of the entity that cannot be modified/deleted.
+            entity_id: The ID of the entity.
+            dependent_entities: A list of entities that depend on this one.
+            status_code: The HTTP status code.
+            error_code: The custom error code.
+
+        Returns:
+            A MockResponse object representing the dependency constraint error.
+        """
         message = f"Cannot perform operation on {entity_type} '{entity_id}' due to existing dependencies"
 
         return MockResponse(
@@ -217,6 +333,21 @@ class BusinessLogicConstraintBuilder:
         error_code: str = "BUSINESS_RULE_VIOLATION",
         details_function: Optional[Callable[[Dict[str, Any]], Dict[str, Any]]] = None,
     ) -> Callable[[Dict[str, Any]], Optional[MockResponse]]:
+        """
+        Creates a validator function based on a custom business rule.
+
+        Args:
+            rule_name: The name of the business rule.
+            validator_function: A function that takes data and returns True if the rule passes, False otherwise.
+            error_message: The error message to use if the rule fails.
+            status_code: The status code for the error response.
+            error_code: The error code for the error response.
+            details_function: An optional function to generate details for the error response.
+
+        Returns:
+            A function that takes request data and returns a MockResponse if the rule fails, else None.
+        """
+
         def rule_validator(data: Dict[str, Any]) -> Optional[MockResponse]:
             if not validator_function(data):
                 details = None
